\name{climatol-internal}
\alias{depudm}
\alias{matpesos}
\alias{stdrz}
\alias{datest}
\alias{snhtw}
\alias{snht}
\alias{estelim}
\alias{plotstan}
\alias{cerrar}
\alias{anom}
\alias{anyf}
\alias{anyi}
\alias{dah}
\alias{dahs}
\alias{dat}
\alias{dat.c}
\alias{dat.d}
\alias{dat.e}
\alias{dat.m}
\alias{dat.m0}
\alias{datmed}
\alias{dat.na}
\alias{dat.s}
\alias{dat.s0}
\alias{dat.z}
\alias{est.c}
\alias{est.d}
\alias{est.p}
\alias{est.w}
\alias{iest}
\alias{mindist}
\alias{na}
\alias{nd}
\alias{ne}
\alias{oneref}
\alias{outan}
\alias{refhom}
\alias{refmed}
\alias{refstd}
\alias{sanom}
\alias{used}
\alias{verde}
\title{Internal climatol objects}
\description{
  Internal climatol objects 
}
\details{
  These objects are used internally and not intended to be
  directly called by the user.
}
\keyword{ internal }
