\name{datcli}
\docType{data}
\alias{datcli}
\title{Monthly climatic data of a single station}
\description{
  Monthly climatic data, for a single station, of average precipitation,
  average daily maxima and minima temperatures, and daily minima temperatures.
}
\usage{data(datcli)}
\format{A matrix of 4 lines (variables) and 12 columns (months) of data}
\source{Made up data for the example of \code{\link{diagwl}}}
\keyword{datasets}
