\name{windfr}
\docType{data}
\alias{windfr}
\title{Wind frequency table}
\description{
  Wind frequency table for rosavent.R example.
}
\usage{data(windfr)}
\format{A table of 4 lines (speed intervals) and 16 columns (wind directions)
  of data, plus headers.}
\source{Invented data for the example of \code{\link{rosavent}}}
\keyword{datasets}
