\name{windfreq.dat}
\docType{data}
\alias{windfreq.dat}
\title{Wind frequency table}
\description{
  Wind frequency table for rosavent.R example.
}
\usage{data(windfreq.dat)}
\format{A table of 4 lines (speed intervals) and 16 columns (speed directions)
  of data, plus headers.}
\source{Invented data for the example of \code{\link{rosavent}}}
\keyword{datasets}
