\name{climatol-internal}
\alias{depudm}
\alias{matpesos}
\alias{stdrz}
\alias{datest}
\alias{snhtw}
\alias{snht}
\alias{estelim}
\alias{plotstan}
\alias{cerrar}
\title{Internal climatol functions}
\description{
  Internal climatol functions
}
\details{
  These functions are used internally and not intended to be
  directly called by the user.
}
\keyword{ internal }
