\name{climatol-internal}
\alias{cerrar}
\alias{climatol.version}
\alias{datmed.mean}
\alias{datmed.sd}
\alias{snht}
\alias{snhtw}
\alias{dat}
\alias{est.c}
\title{Internal climatol functions}
\description{
  Internal climatol functions
}
\details{
  These functions are used internally and not intended to be
  called directly by the user.
}
\keyword{internal}
