\name{climatol-internal}
\alias{cerrar}
\alias{climatol.version}
\alias{cuct}
\alias{read.dat}
\alias{snht}
\alias{wtest}
\alias{unsufix}
\title{Internal \code{climatol} functions}
\description{
  Internal \code{climatol} functions
}
\details{
  These functions are used internally and are not intended to be
  called directly by the user.
}
\keyword{misc}
