\name{datcli}
\alias{datcli}
\docType{data}
\title{Monthly climatic data of a single station}
\description{
  Monthly climatic data, for a single station, of average precipitation,
  average daily maximum and minimum temperatures, and daily minimum
  temperatures.
}
\usage{data(datcli)}
\format{A matrix of 4 lines (variables) and 12 columns (months) of data}
\source{Made up data for the example of \code{\link{diagwl}}}
\seealso{\code{\link{diagwl}}}
\examples{ 
data(datcli) 
print(datcli) 
}
\keyword{datasets}
