\name{windfr}
\alias{windfr}
\docType{data}
\title{Wind frequency table}
\description{
  Example of wind frequency contingency table.
}
\usage{data(windfr)}
\format{A table of 4 lines (speed intervals) and 16 columns (wind directions)
  of data, plus headers.}
\source{Bogus data for the example of \code{\link{rosavent}}}
\seealso{\code{\link{rosavent}}}
\examples{
data(windfr)
print(windfr)
}
\keyword{datasets}
